modules creation and tested
