+++ index.html
@@ -9,6 +9,8 @@
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-GLhlTQ8iRABdZLl6O3oVMWSktQOp6b7In1Zl3/Jr59b6EGGoI1aFkw7cmDA6j6gD" crossorigin="anonymous">
     <script src="scripts.js" defer></script>
+    <script src="modules/components/navBar.js" defer></script>
+    <script src="modules/components/body.js" defer type="module"></script>
 </head>
 
 <body>
