+++ modules/components/todoList.js
@@ -1,0 +1,19 @@
+export const todoList = `<div class="card mb-5" style="width: 18rem;">
+<div class="card-body">
+    <div class="mb-3">
+        <input type="title" class="form-control" name="title" placeholder="TODO">
+    </div>
+    <div class="mb-3">
+        <input type="date" class="form-control" name="date" placeholder="">
+    </div>
+    <div class="mb-3">
+        <textarea class="form-control" name="description" rows="3"></textarea>
+    </div>
+    <div class="mb-3">
+        <div class="btn-group w-100" role="group" aria-label="Basic mixed styles example">
+            <button type="button" class="btn btn-danger">URGENTE</button>
+            <button type="button" class="btn btn-success">Hecho</button>
+        </div>
+    </div>
+</div>
+</div>`
